Run the unu command line tool

Run the unu command line tool

unu(cmd = "", args = NULL, DryRun = FALSE, ...)

find_unu(location = getOption("runu.unu"))

Arguments

cmd
The name of the unu command to run

args
Additional arguments to unu

DryRun
When TRUE show what would happen rather running command

...
Additional arguments passed to system

location
- location of the unu executable

Value

unu returns the value of system after running unu.

Details

If find_unu cannot find your unu executable then either add the directory to your path or set the package option.

Installation

You will normally need to compile unu, but it is cross-platform (Windows/Mac/Linux) and has minimal dependencies. It is installed as part of the teem libraries available at http://teem.sourceforge.net/. See in particular the download and build pages for details.

Examples

# NOT RUN {
# manually set unu location if required
options(runu.unu="/path/to/unu")
# }